refactor(numerical-semigroups): partition summary contracts
docs(numerical-semigroups): publish summary admission bounds
